Drawer, order and style changes
- Drawer now toggles when coming back from pages that are accessed through the drawer itself

- Added title of restaurant to orders page

- If connected to a restaurant the restaurant item will change styles to indicate you are connected to that restaurant

- Created StyleSheet for order.js
