+++ components/orderHistory.js
@@ -72,7 +72,10 @@ export default class OrderHistory extends Component {
         <SafeAreaView style={navStyles.navBar}>
           <TouchableOpacity
             style={navStyles.navTab}
-            onPress={() => this.props.navigation.goBack()}
+            onPress={() => {
+              this.props.navigation.goBack();
+              this.props.navigation.toggleDrawer();
+            }}
           >
             <ArrowBack width={35} height={35} />
           </TouchableOpacity>
